feat(source-control): add header sync actions
