+++ src/modules/source-control/index.ts
@@ -1,1 +1,6 @@
 export { SourceControlPanel } from "./SourceControlPanel";
+export {
+  getSourceControlRemoteIndicator,
+  useSourceControl,
+  type SourceControlSummary,
+} from "./useSourceControl";
